Add some convenient tools to the vagrant machine when it boots

+++ custom/installZSH.sh
@@ -60,8 +60,11 @@ install_zsh() {
   # Copy snippets
   echo "Copy snippets to vagrant user..."
   su -l vagrant -s "/bin/sh" -c "mkdir -p ${VAGRANT_HOME}/.config/pet"
-  su -l vagrant -s "/bin/sh" -c "cp ${CUSTOM_PET_SNIPPETS} ${VAGRANT_HOME}/.config/pet/snippet.toml"
+  
   su -l vagrant -s "/bin/sh" -c "cp ${CUSTOM_PET_CONFIG} ${VAGRANT_HOME}/.config/pet/config.toml"
+  if [[ ! -f "${CUSTOM_PET_SNIPPETS}" ]]; then 
+    su -l vagrant -s "/bin/sh" -c "cp ${CUSTOM_PET_SNIPPETS} ${VAGRANT_HOME}/.config/pet/snippet.toml"
+  fi
 }
 
 install_zsh_root() {
@@ -114,7 +117,11 @@ install_zsh_root() {
   # Copy snippets
   echo "Copy snippets to root user..."
   sudo mkdir -p "${ROOT_HOME}"/.config/pet
-  sudo cp "${CUSTOM_PET_SNIPPETS}" "${ROOT_HOME}"/.config/pet/snippet.toml
+  
+  sudo cp "${CUSTOM_PET_CONFIG}" "${ROOT_HOME}/.config/pet/config.toml"
+  if [[ ! -f "${CUSTOM_PET_SNIPPETS}" ]]; then 
+    sudo cp "${CUSTOM_PET_SNIPPETS}" "${ROOT_HOME}"/.config/pet/snippet.toml
+  fi
 }
 
 install_miscellaneous() {
